[Snapshot] in middle of refactoring, collector passing tests
