--- src/bpm_data_combiner/data_model/collection_item.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class CollectionItem:
-    cnt: int
-    name: str
-    payload : object
