+++ src/bpm_data_combiner/sub_pkgs/__init__.py
@@ -1,0 +1,3 @@
+"""packages that could be used stand alone for similar applications
+
+"""
