+++ src_collector/collector/__init__.py
@@ -1,0 +1,19 @@
+"""Combine data from different sources using identifiers
+
+Identifiers are
+
+* name of the sender
+* package id
+
+Shall still work if packages arrive out of order
+"""
+from .bl.collector import Collector
+from .data_model.collection_item import CollectionItem
+from .errors import DoubleSubmissionError, UnknownDeviceNameError
+
+__all__ = [
+    "CollectionItem",
+    "Collector",
+    "DoubleSubmissionError",
+    "UnknownDeviceNameError",
+]
