+++ src_collector/collector/data_model/collection_item.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+from typing import Union, Hashable
+
+
+@dataclass
+class CollectionItem:
+    #: identifier of the package, unique for the sender
+    id_: Union[int|Hashable]
+    #: identifier for the sender
+    name: str
+    payload : object
