+++ src_collector/collector/errors.py
@@ -1,9 +1,9 @@
 class DoubleSubmissionError(AssertionError):
     """Data submitted twice"""
 
+
 class UnknownDeviceNameError(AssertionError):
-    """Device name unknown
-    """
+    """Device name unknown"""
 
 
 __all__ = ["DoubleSubmissionError", "UnknownDeviceNameError"]
