+++ src_collector/collector/interfaces/collection_for_one_id.py
@@ -1,0 +1,9 @@
+"""
+Todo:
+    required? is it accessible from user space ?
+"""
+from abc import ABCMeta, abstractmethod
+
+
+class CollectionForOneIdInterface(metaclass=ABCMeta):
+    pass
