+++ src_collector/pyproject.toml
@@ -1,0 +1,25 @@
+[build-system]
+requires = [ "poetry-core>=1.0.0",]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+name = "collector"
+version = "0.0.1"
+description = "Combine data from different sources using package id"
+authors = [ "Pierre Schnizer"]
+license = "GPL-3.0"
+homepage = "https://github.com/hz-b/bpm-data-combiner"
+documentation = "https://hz-b.github.io/bpm-data-combiner/"
+repository = "https://github.com/hz-b/bpm-data-combiner.git"
+keywords = []
+
+[tool.poetry.dependencies]
+python = ">=3.7"
+numpy = "*"
+
+Sphinx = { version = "*", optional = true }
+sphinx-rtd-theme = { version = "*", optional = true }
+sphinxcontrib-napoleon = { version = "*", optional = true }
+
+[tool.poetry.extras]
+docs = ["Sphinx", "sphinx-rtd-theme", "sphinxcontrib-napoleon"]
